frontend/requirements.txt
streamlit
python-socketio[asyncio_redis]
aiohttp